Move logic to create events
